build: Always use neon on arm, remove support for arm_version=6
As far as I know, the last client of arm_use_neon=false was remove a
while ago, so remove having a toggle for it.

Similarly for arm_version 6.

Shout on the review if this breaks some use case I'm not aware of!

Bug: 469092869
Change-Id: I6c201c13e72a69f1efbbfac8d128f71eaca78934
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7262201
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559405}
NOKEYCHECK=True
GitOrigin-RevId: e6b1ab272575ae1bebae942dd3b2368d26a51b14