build: Remove arm_tune gn arg
As far as I can tell, nothing ever overrides this.

This does have the effect that we no longer pass -mtune=generic-armv8-a
in arm64 builds. I'm relying on the perf bots to confirm that dropping
that flag is as harmless as it looks.

Bug: 469092869
Change-Id: I2fb22902fb011882e2d84ebeb1ba2f97f0d59ca4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7265933
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1559979}
NOKEYCHECK=True
GitOrigin-RevId: 15ba1ec6c7b87ddf8621758ee1f5b3d809a265ab