Allow circular includes in cxx targets
This change forwards the `allow_circular_includes_from` GN option to
generated `cxx` targets, allowing C++ shims to include files from them
without introducing circular dependencies. This is important for writing
shims with functions that e.g. take in both a C++ type and a Rust type
and are exposed to Rust, since those files need to both depend on the
bindings (so the Rust type is declared) and export the function to the
bindings (so Rust can call it).

Bug: 469057512
Change-Id: I1c3acfb3b32b67d9a1ac9b1cbb012b6d8efd6f64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7270085
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#1560155}
NOKEYCHECK=True
GitOrigin-RevId: 9d5b6866f22ba6841a6ee9be066a75bf9fce4bd0